Fix redis startup in test

+++ test/acpc_table_manager_exe_test.rb
@@ -21,7 +21,11 @@ def my_setup
   File.open(config_file, 'w') do |f|
     f.puts YAML.dump(CONFIG_DATA)
   end
-  redis_pid = Process.spawn('redis-server')
+  redis_pid = if AcpcTableManager.new_redis_connection.connected?
+    nil
+  else
+    Process.spawn('redis-server')
+  end
   patient_pid = Process.spawn(
     "#{File.join(PWD, '..', 'exe', 'acpc_table_manager')} -t #{config_file}"
   )
@@ -31,7 +35,7 @@ end
 tmp_dir, config_file, redis_pid, patient_pid = my_setup
 
 def my_teardown(tmp_dir, config_file, redis_pid, patient_pid)
-  AcpcDealer.kill_process(redis_pid)
+  AcpcDealer.kill_process(redis_pid) if redis_pid
   AcpcDealer.kill_process(patient_pid)
   FileUtils.rm_rf tmp_dir
   Timeout.timeout(10) do
